requirements.txt
Django
djangorestframework
markdown
django-filter
channels
bitmex
websocket_client